pyproject.toml
[project]
name = "pngalpha"
version = "0.1.0"
description = "Two-pass alpha extraction tool for creating transparent PNG images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "tkinterdnd2>=0.4.2",
]

[project.scripts]
pngalpha = "pngalpha.cli:main"
pngalpha-gui = "pngalpha.gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pngalpha"]
